Move static methods from class EngineWrapper to few separate files.
My God, doesn't this look so much better now...
