+++ optimization_engine/engine.py
@@ -10,6 +10,7 @@ from typing import Tuple, List, Literal, Any, Union, Callable
 import itertools
 from numpy import dtype, ndarray
 from . import constants as cn
+from . import metrics, random_generators, utils, genetic
 
 
 class EngineWrapper:
@@ -42,74 +43,17 @@ class EngineWrapper:
         self.rewards_matrix = None
         self.coordinates = None
 
-    @staticmethod
-    def generate_random_city_cords(cities_amount:np.uint8, max_city_size = np.uint32(2**16 - 1),
-                                   max_coordinate_val = np.uint32(2**16 - 1), seed: int = None) \
-            -> Tuple[np.ndarray[np.uint32], np.ndarray[np.uint32]]:
-        """
-        This function generates random cities coordinates on a square map and random city sizes. Returns results as two
-        separate ndarrays. Providing the seed allows to reproduce results. This function guarantees, that no 2 cities
-        will have the exact same coordinates
-
-        Coordinates are returned as 2D ndarray of shape (cities_amount, 2) and sizes are returned as 1D ndarray of shape
-        (cities_amount,)
-
-        :param max_coordinate_val: How big is side of the map. Smallest possible map is 32 x 32, biggest possible map
-            is (2^31 - 1) x (2^31 - 1). Not 2^32 to guarantee that the distance does not overflow
-        :param max_city_size: How big can be a single city. This value cannot be smaller than 64 and can go up to 2^32-1
-        :return: A tuple: (Coordinates, sizes_vector)
-        """
-        if cities_amount < 2 or cities_amount > 255:
-            wrong_cities_amount_message = (f"Cities amount expected to be a number between 2 and 255 (inclusive), "
-                                           f"got {cities_amount} instead")
-            raise ValueError(wrong_cities_amount_message)
-        if max_coordinate_val < 32 or max_coordinate_val > 2**31 - 1:
-            wrong_coordinate_val = (f"max_coordinate_val expected to be a number between 32 and 2^31 - 1 (inclusive), "
-                                    f"got {max_coordinate_val} instead")
-            raise ValueError(wrong_coordinate_val)
-        rng = np.random.default_rng(seed)
-        max_city_size = np.uint32(64) if max_city_size < 64 else max_city_size
-
-        coordinates_list = []
-        already_used_coordinates = set()
-        while len(coordinates_list) < cities_amount:
-            one_pair = rng.integers(max_coordinate_val, size=2, dtype=np.uint32)
-            if tuple(one_pair) not in already_used_coordinates:
-                coordinates_list.append(one_pair)
-                already_used_coordinates.add(tuple(one_pair))
-        coordinates = np.array(coordinates_list, dtype=np.uint32)
-        cities_sizes = rng.integers(max_city_size, size=(cities_amount,), dtype=np.uint32)
-        # noinspection PyTypeChecker
-        return coordinates, cities_sizes
-
-    @staticmethod
-    def calculate_distances_matrix(coordinates: ndarray[np.uint32], use_manhattan_metric=True) -> ndarray[np.uint32]:
-        """
-        this method takes a 2D coordinates array of shape (cities_amount, 2) and returns a 2D ndarray of shape
-        (cities_amount, cities_amount) where element with index [i,j] is the distance between cities i and j.
-        :param use_manhattan_metric: If set to True (default) evaluates Manhattan distance - sum of the absolute
-            differences between the coordinates. When set to False standard Euclidean metric will be used (rounding up
-            to a whole integer)
-        """
-        if coordinates.ndim != 2 or coordinates.shape[1] != 2:
-            raise ValueError(f"Expected coordinates to have shape (cities_amount, 2), got {coordinates.shape} instead")
-        cities_amount = len(coordinates)
-        distances_matrix = np.zeros(shape=(cities_amount, cities_amount), dtype=np.uint32)
-        cords = coordinates.astype(np.int64)  # changing to int64 to allow for negative differences and to 'fit'
-        # squaring when calculating Euclidean
-        # This loop will leave diagonal not-computed (initialized with 0). It is intentional since distance from a
-        # city to itself is always 0
-        for row in range(cities_amount):
-            for col in range(row):
-                difference = cords[row] - cords[col]  # so that it can be negative
-                if use_manhattan_metric:
-                    distances_matrix[row, col] = np.sum(np.abs(difference))
-                else:
-                    difference = np.sqrt(np.sum(np.square(difference)))
-                    distances_matrix[row, col] = np.round(difference)
-        distances_matrix = EngineWrapper.symmetrize_numpy_matrix(distances_matrix, "lower")  # have to copy lower side
-        # because the lower side was computed
-        return distances_matrix
+    function_F = staticmethod(metrics.function_F)
+    calculate_distances_matrix = staticmethod(metrics.calculate_distances_matrix)
+    calculate_reward_matrix = staticmethod(metrics.calculate_reward_matrix)
+    goal_achievement_function = staticmethod(metrics.goal_achievement_function)
+    symmetrize_numpy_matrix = staticmethod(utils.symmetrize_numpy_matrix)
+    random_choose_2_arrays = staticmethod(utils.random_choose_2_arrays)
+    generate_random_city_cords = staticmethod(random_generators.generate_random_city_cords)
+    generate_one_candidate = staticmethod(random_generators.generate_one_candidate)
+    select_parents_tournament = staticmethod(genetic.select_parents_tournament)
+    mutate_bool_ndarray = staticmethod(genetic.mutate_bool_ndarray)
+    crossover = staticmethod(genetic.crossover)
 
     def generate_cities(self, cities_amount:np.uint8 = None, max_city_size: np.uint32 = None,
                         max_distance: np.uint32 = None, seed: int = None) \
@@ -130,174 +74,6 @@ class EngineWrapper:
         coordinates, sizes_vector = self.generate_random_city_cords(cities_amount, max_city_size, max_distance, seed)
         distances_matrix = self.calculate_distances_matrix(coordinates, use_manhattan_metric=True)
         return distances_matrix, sizes_vector, coordinates
-
-    # noinspection PyIncorrectDocstring
-    @staticmethod
-    def function_F(size1: np.uint32, size2: np.uint32, distance: np.uint32, max_result_val: np.uint64 = np.uint64(5e10)) \
-            -> np.uint64:
-        """
-        this function takes two city sizes and a distance between the cities and returns the amount of money earned
-        for connecting those two together. By definition F(sizeA, sizeB, dist) = F(sizeB, sizeA, dist).
-        With uint64 there is no risk of an overflow when using this function, even with max (2**32 - 1) possible values
-        as inputs. If inputs are larger (they never should be) they will be truncated to numpy.iinfo(np.uint32).max
-
-        This function implements tanh(distance/(10^6)) * max_result_val * (ln(size1) + ln(size2)). It is always increasing
-        with respect to all of its inputs. However, tanh is scaled in such a way that for distances smaller than 100 000
-        it is almost linear. For distance of one million it achieves around 76% of given max_result_val, with distance of
-        2 million it is 96%. Increasing distances greater than 100 000 start to yield noticeably lower money increase
-        (sigmoid-like shape)
-
-        The above formula is purely arbitrary and any callable taking two sizes and a distance that returns
-        one numpy.uint64 can be used instead. If you want to be sure nothing overflows (neither numpy nor this program will
-        warn you about it!!) make sure that value returned by your custom callable is never bigger than 5.6e14
-        since max money value is  1.844674e+19 and as for this moment max 32 385 connections are possible (255 choose 2).
-        Of course this will not be correct if number of cities is greater than 255 (which at the moment of writing this doc
-        is not allowed and there are no plans to change it in the future)
-
-        :param max_result_val: this parameter represents the supremum of the whole formula - this value is a limit of
-            this expression as all 3 inputs go to +infinity. For maximum possible (2^^32-1) inputs computed value is so
-            close to this parameter that a difference is too small to be represented in np.64float
-        """
-        max_param_val = np.iinfo(np.uint32).max
-        size1 = size1 if size1 < max_param_val else max_param_val
-        size2 = size2 if size2 < max_param_val else max_param_val
-        distance = distance if distance < max_param_val else max_param_val
-
-        # this performs a tanh(distance) * log(size1*size2) function with supremum of max_result_val
-        money = (np.log(size1) + np.log(size2))/(2 * np.log(max_param_val)) * np.tanh(distance/(10**6)) * max_result_val
-        return money.astype(np.uint64)  # there is absolutely no point in being more precise than an
-        # int value, especially since in this case "money" usually is in magnitude of millions or larger
-
-    @staticmethod
-    def calculate_reward_matrix(distances_matrix: ndarray[np.uint32], sizes_vector: ndarray[np.uint32],
-                                achievement_function: Callable = None, max_result_val: np.uint64 = None)\
-            -> np.ndarray[np.uint64]:
-        """
-        This method allows to leverage the fact, that in the solved optimization problem 'alleles' are both binary
-        and completely independent. This means one can calculate how much 'reward' each individual allele brings and
-        avoid calculating this each time. Of course this 'reward_matrix' can be used only when distances between cities,
-        their sizes and the function used to calculate reward remain CONSTANT. When they change, the reward matrix has
-        to be calculated once again for new parameters.
-        :param distances_matrix: A symmetric 2D numpy array representing distances between cities
-        :param sizes_vector: A 1D numpy array representing the size of each city
-        :param achievement_function: A Callable that accepts the same parameters as
-            EngineWrapper.function_F(). When not provided, the default function_F() will be used. Refer to docs of
-            EngineWrapper.function_F() for details about customizing the reward function.
-        :param max_result_val: This parameter allows to specify the reward function's value for max possible inputs.
-            If None, it will NOT be passed to the reward function at all - your custom function does not have to take it
-        :return: A reward matrix - 2D np.ndarray of the same shape as `dist`. Element with index [i,j] represents how
-            much money does one earn by connecting cities with indexes i and j. Of course this matrix by definition
-            is symmetric.
-        """
-        achievement_function = EngineWrapper.function_F if achievement_function is None else achievement_function
-        if distances_matrix.ndim !=2 or sizes_vector.ndim != 1:
-            dim_mismatch_err_message = (f"Expected distances matrix to have 2 dimensions and sizes vector to have one,"
-                                        f"but received distances with {distances_matrix.ndim} dimensions and sizes with"
-                                        f" {sizes_vector.ndim} dimensions instead")
-            raise ValueError(dim_mismatch_err_message)
-        if np.any(distances_matrix != distances_matrix.T):
-            raise ValueError("Expected distances_matrix to be symmetric, not symmetric matrix received")
-
-        # noinspection PyTypeChecker
-        rewards_matrix: ndarray[np.uint64] = np.zeros_like(distances_matrix, dtype=np.uint64)
-
-        if max_result_val is None:
-            for row_no in range(rewards_matrix.shape[0]):
-                for col_no in range(row_no):
-                    rewards_matrix[row_no, col_no] = achievement_function(sizes_vector[row_no], sizes_vector[col_no],
-                                                                          distances_matrix[row_no, col_no])
-        else:
-            for row_no in range(rewards_matrix.shape[0]):
-                for col_no in range(row_no):
-                    rewards_matrix[row_no, col_no] = achievement_function(sizes_vector[row_no], sizes_vector[col_no],
-                                                                      distances_matrix[row_no, col_no], max_result_val)
-
-        rewards_matrix = EngineWrapper.symmetrize_numpy_matrix(rewards_matrix, "lower")  # watch out, lower
-        # side is filled so lower has to be copied!
-        return rewards_matrix
-
-    # noinspection PyIncorrectDocstring
-    @staticmethod
-    def goal_achievement_function(distances_matrix: np.ndarray[np.uint32], sizes_vector: np.ndarray[np.uint32],
-                                  connections_matrix: np.ndarray[bool], max_cost: np.uint64,  max_railway_len: np.uint64,
-                                  max_connections_count: np.uint32, one_rail_cost: np.uint32,
-                                  one_infrastructure_cost: np.uint32, rewards_matrix: ndarray[np.uint64] = None) \
-            -> np.uint64:
-        """This function takes exactly one solution (or solution candidate) and evaluates it with non-negative score.
-        On the caller of this function lies the responsibility for ensuring that both matrices and the size_vector have
-        appropriate shapes (the same number of cities), that all provided values have correct datatypes and that the
-        matrices are symmetrical (they need to be because of what they represent)
-        :param distances_matrix: Matrix where element m[i][j] is the distance between city i and city j. By definition
-            m[i][j] = m[j][i] (this matrix has to be symmetrical)
-        :param connections_matrix: Matrix of booleans where element m[i][j] represents whether there is a connection
-            between cities i and j. Of course m[i][j] has to be equal to m[j][i], same as with the distances_matrix.
-        :param rewards_matrix: 2D ndarray of shape (cities_amount, cities_amount) - the same shape as distances_matrix.
-            If provided, element with index [i,j] will represent the reward for connecting cities with indexes i and j.
-            If in your optimization problem distances, sizes and function used to calculate reward are constant
-            do not change often calculating this matrix beforehand is HIGHLY RECOMMENDED - it will save a ton of
-            computation and improve performance.
-         :returns: If the solution does not satisfy any of the constraints (for example too many connections) this function
-             returns exactly 0. Otherwise, it returns one positive integer representing a value of this solution -
-             the bigger, the better
-        """
-        if distances_matrix.ndim != 2 or distances_matrix.shape[0] != distances_matrix.shape[1]:
-            raise ValueError(f"distances_matrix should be a 2-dimensional square numpy ndarray, got shape "
-                             f"({distances_matrix.shape}) instead")
-        if sizes_vector.ndim != 1:
-            raise ValueError(f"sizes_vector should be a 1D numpy ndarray, instead got shape ({sizes_vector.shape})")
-        if sizes_vector.shape[0] != distances_matrix.shape[0]:
-            err_shapes_mismatch = f"Distances_matrix and sizes_vector do not have matching shapes, received "\
-                                f"distances_matrix: ({distances_matrix.shape}) and sizes_vector: ({sizes_vector.shape})"
-            raise ValueError(err_shapes_mismatch)
-        if connections_matrix.shape != distances_matrix.shape:
-            err_shape_mismatch = (f"shape of connections_matrix: {connections_matrix.shape} does not match with the "
-                                  f"shape of distances_matrix: {distances_matrix.shape}")
-            raise ValueError(err_shape_mismatch)
-        if rewards_matrix is not None and rewards_matrix.shape != distances_matrix.shape:
-            err_shape_mismatch = (f"Expected the reward matrix to have the same shape as connections matrix, but "
-                                  f"received arrays with shapes {rewards_matrix.shape} and {connections_matrix.shape} "
-                                  f"respectively")
-            raise ValueError(err_shape_mismatch)
-        # first checks for max number of connections, as it requires basically no computation
-        connections_count = np.uint32(np.count_nonzero(connections_matrix)) // 2
-        # needs to be divided by 2 due to the nature of the connections' matrix. Thanks to numpy vectorization looping over
-        # the whole matrix and dividing the result by 2 is way faster than iterating over only half of the matrix
-        if connections_count > max_connections_count:
-            return np.uint64(0)
-
-        # now checks for max railway length
-        built_rails_matrix = np.multiply(connections_matrix, distances_matrix)
-        built_rails_count = np.sum(built_rails_matrix, dtype=np.uint64) // 2  # as with connections result is divided by 2
-        if built_rails_count > max_railway_len:
-            return np.uint64(0)
-
-        # checks for max budget
-        rails_cost = built_rails_count * one_rail_cost
-        # using np.multiply to cast to np.uint64 to prevent any overflows
-        infrastructure_costs = np.multiply(connections_count, one_infrastructure_cost, dtype=np.uint64)
-        total_cost = rails_cost + infrastructure_costs
-        if total_cost > max_cost:
-            return np.uint64(0)
-
-        # if the function did not return up to this point all the additional requirements are met and the actual score can
-        # be calculated. Unfortunately I did not find a way to vectorize it using numpy built-ins and have to
-        # operate manually using indexes
-
-        upper_connections_half = np.triu(connections_matrix)  # this matrix by definition has to be symmetrical so we can
-        # take only one half without any loss of information instead of dividing by 2 at the end - this will save computing
-        rows, cols = upper_connections_half.nonzero()  # extracts indices of cities that are connected
-        connection_indexes = zip(rows, cols)  # each element is a pair of indexes representing one connection
-
-        individual_scores_matrix = np.zeros_like(connections_matrix, dtype=np.uint64)
-        if rewards_matrix is None:
-            for row, col in connection_indexes:
-                individual_scores_matrix[row,col] = EngineWrapper.function_F(sizes_vector[row], sizes_vector[col],
-                                                                             distances_matrix[row,col])
-        else:
-            for row, col in connection_indexes:
-                individual_scores_matrix[row,col] = rewards_matrix[row,col]
-        overall_score = np.sum(individual_scores_matrix)
-        return overall_score
 
     def goal_function_convenient(self, distances_matrix: np.ndarray[np.uint32], sizes_vector: np.ndarray[np.uint32],
                                 connections_matrix: np.ndarray[bool], max_cost: np.uint64 = None,
@@ -355,35 +131,6 @@ class EngineWrapper:
 
             # noinspection PyTypeChecker
             return resulting_arr
-
-
-    @staticmethod
-    def symmetrize_numpy_matrix(matrix: np.ndarray, copied_side: Literal["upper","lower"] = "upper") -> np.ndarray|None:
-        """
-        this method takes one matrix and returns symmetric matrix - when copied_side is 'upper' then upper side is
-        copied to the lower side, analogically with copied_side = 'lower'. If provided array has more than 2 dimensions
-        the operation is applied to the final two axes. If provided array has less than 2D None is returned
-        """
-        if matrix.ndim < 2:
-            raise ValueError(f"An array to symmetrize needs to have at least 2 dimensions, got {matrix.ndim} instead")
-        if matrix.shape[-1] != matrix.shape[-2]:
-            raise ValueError(f"Final two dimensions need to have the same size, got ({matrix.shape[-1]}, {matrix.shape[-2]}) instead")
-        if copied_side == "upper":
-            return np.triu(matrix, k=1).swapaxes(-1, -2) + np.triu(matrix, k=0)
-        if copied_side == "lower":
-            return np.tril(matrix, k=-1).swapaxes(-1, -2) + np.tril(matrix, k=0)
-
-    @staticmethod
-    def generate_one_candidate(rng: np.random.Generator, cities_amount) -> np.ndarray[Any, dtype[bool]]:
-        """This method returns one candidate, i.e. one connection matrix with shape (cities_amount, cities_amount).
-        This matrix is symmetric, and it's diagonal is always False (since a city cannot be connected to itself)"""
-
-        rng = rng if rng is not None else np.random.default_rng()
-        one_candidate = rng.integers(0, 1, size=(cities_amount, cities_amount), dtype=bool, endpoint=True)
-        one_candidate = EngineWrapper.symmetrize_numpy_matrix(one_candidate, "upper")
-        np.fill_diagonal(one_candidate, False)
-        return one_candidate
-
 
     def _check_one_candidate(self, one_candidate: np.ndarray[bool],
                              distances_matrix: np.ndarray[np.uint32], sizes_vector: np.ndarray[np.uint32],
@@ -445,120 +192,6 @@ class EngineWrapper:
             population[i] = new_elem
         # noinspection PyTypeChecker
         return population
-
-    @staticmethod
-    def select_parents_tournament(goal_scores: np.array, one_fight_size: int = 3,
-                                  excluded_candidates: int = 3, rng: np.random.Generator = None) -> ndarray | None:
-        """
-        This method takes a 1D numpy ndarray with scores and performs tournament selection, then returns indexes of
-            elements from the scores array that 'passed' the tournament
-        :param goal_scores: 1D numpy array containing scores. Will return None if the array is not 1D
-        :param one_fight_size: How many solutions are in one tournament group. Only the best one from each group
-            will be selected, others are forgotten forever...
-        :param excluded_candidates: How many best solutions are excluded from the tournament. These candidates do not
-            take part in the tournament at all, meaning no solution will be 'killed' simply because it was worse in
-            comparison.
-        :param rng: Numpy random.Generator instance. It is used to 'randomly' divide solutions into tournament groups.
-            Passing identical instance of the generator (WITH EXACTLY THE SAME STATE - watch for earlier uses of passed
-            rng!) will ensure reproducibility.
-        :return: Numpy ndarray containing indexes of chosen parents. Indexing the array used to create `goal_scores`
-            with this returned array will give the chosen solutions
-        """
-        if goal_scores.ndim != 1:
-            return None
-        rng = np.random.default_rng() if rng is None else rng
-        # sorted_scores_idx = np.argsort(goal_scores)[::-1]  # oddly specific comment about the need to reverse the
-        # argsort()'s result since it is sorting in ascending order...
-        sorted_scores_idx = np.argpartition(goal_scores, kth=-excluded_candidates)[::-1]
-        winners = []
-        for i in range(excluded_candidates):
-            winners.append(sorted_scores_idx[i])
-        indexes = rng.permutation(np.arange(start=excluded_candidates, stop=len(goal_scores)))
-        # indexes = np.array_split(np.array(indexes),
-        #                          indices_or_sections=(len(goal_scores) - guaranteed_survivors)// tournament_size + 1)
-        tournament_scores = [goal_scores[i] for i in indexes]
-
-        indexes = np.array_split(np.array(indexes),
-                                 indices_or_sections=(len(goal_scores) - excluded_candidates) // one_fight_size + 1)
-
-        tournament_scores = np.array_split(np.array(tournament_scores),
-                                           indices_or_sections=(len(goal_scores) - excluded_candidates) // one_fight_size + 1)
-
-        tournament_winners_positions = [np.argmax(one_fight) for one_fight in tournament_scores]
-        for fight_no, winner in enumerate(tournament_winners_positions):
-            winners.append(indexes[fight_no][winner])
-
-        return np.array(winners)
-
-    @staticmethod
-    def random_choose_2_arrays(a: np.ndarray, b: np.ndarray, rng: np.random.Generator = None) -> np.ndarray:
-        """Takes two ndarrays of the same shape and returns another ndarray of the same shape that has
-        its elements randomly chosen from `a` and `b`"""
-        if a.shape != b.shape:
-            raise ValueError(f"Random choose from 2 arrays expected the arrays to have the same shape, got {a.shape} and {b.shape} instead")
-        rng = rng if rng is not None else np.random.default_rng()
-        child_indexes = rng.integers(0, 1, size=a.shape, endpoint=True)
-        child = np.where(child_indexes == 0, a, b)
-        return child
-
-    @staticmethod
-    def crossover(parents: np.ndarray, offspring_count: int, rng: np.random.Generator = None)\
-            -> np.ndarray:
-        """
-        This function takes a 3D numpy ndarray with shape (parents_count, (...)) and returns another 3D ndarray of shape
-        (parents_count + offspring_count, (...)) in which first parents_count entries are just copied from `parents`
-        and every next entry is a 'child' - an array with elements randomly chosen from two random 'parents'.
-
-        :param parents: Numpy ndarray containing exactly 3 dimensions where the first (i.e. shape[0]) dimension represents
-            the batch size
-        :param offspring_count: How many additional children will be created
-        :param rng: If provided the output of this function is fully reproducible
-        """
-        if parents.ndim != 3:  # batch of 2D connection matrices is expected
-            raise ValueError(f"Crossover expected a np.ndarray with exactly 3 dimensions, got {parents.ndim} instead")
-        parents_amount = parents.shape[0]
-        offspring = np.empty((parents_amount + offspring_count, parents.shape[1], parents.shape[2]), dtype=parents.dtype)
-        offspring[0:parents_amount, :, :] = parents
-        rng = np.random.default_rng() if rng is None else rng
-        for i in range(offspring_count):
-            indexes = rng.choice(len(parents), size=(2,), replace=False)
-            chosen_parents = parents[indexes]
-            one_child = EngineWrapper.random_choose_2_arrays(chosen_parents[0], chosen_parents[1], rng)
-            offspring[i + parents_amount, :, :] = one_child
-        return offspring
-
-    @staticmethod
-    def mutate_bool_ndarray(arr: ndarray[Any, dtype[bool]], mutation_chance = 1e-2,rng: np.random.Generator = None,
-                            spare_indexes: ndarray = None, create_copy: bool = False) -> ndarray[Any, dtype[bool]]:
-        """
-        This method takes a numpy ndarray of any shape and of bool datatype. Each element in this array has
-        a `mutation_chance` chance to be logically flipped. Optionally `spare_indexes` can be provided to locally
-        prevent mutations, which is necessary to establish elitism. If `create_copy` is provided this function will
-         not make any modifications on the original and return a new array instead.
-
-        :param spare_indexes: If boolean - Ndarray with exactly the same shape as `arr`. If an element of this array is
-            True then the value with the same index in `arr` is guaranteed to NOT be mutated. If integers - 1D ndarray
-            containing indexes which will NOT be mutated. Integers allow only to index the 0-th dimension - if more
-            precision is required, use an array of booleans
-            """
-
-        if spare_indexes is not None and isinstance(spare_indexes.dtype, bool) and spare_indexes.shape != arr.shape:
-            err_shape_mismatch_message = (f"boolean spare_indexes need to have the same shape as `arr`, but received "
-                                          f"shapes {arr.shape} and {spare_indexes.shape}")
-            raise ValueError(err_shape_mismatch_message)
-        if spare_indexes is not None and np.issubdtype(spare_indexes.dtype, np.integer) and spare_indexes.ndim != 1:
-            err_dim_mismatch_message = (f"integer spare_indexes need to be exactly one-dimensional, but an array with"
-                                        f"{spare_indexes.ndim} dimensions was provided")
-            raise ValueError(err_dim_mismatch_message)
-
-        new_arr = arr.copy() if create_copy else arr
-
-        rng = rng if rng is not None else np.random.default_rng()
-        flipped_mask = rng.choice([True, False], size=arr.shape, p=[mutation_chance, 1 - mutation_chance])
-        if spare_indexes is not None:
-            flipped_mask[spare_indexes] = False
-        new_arr[flipped_mask] = np.logical_not(new_arr[flipped_mask])
-        return new_arr
 
     def genetic_algorithm(self, distances_matrix: np.ndarray[np.uint32], sizes_vector: np.ndarray[np.uint32],
                           initial_population: np.ndarray[bool] = None, population_size = 100, seed = None,
@@ -655,5 +288,3 @@ def simple_test():
             for solution in last_population.astype(np.ushort):  # changed to integer type for more concise printing
                 print(solution)
                 print("\n")
-
-
